Added function to FaceDefinition: addDefaultFacePose() which allows you to set various blends on the face and those effectively become the 'neutral face'. For example, you can add a small smile to the character, a furrowed brow, etc. and that configuration will be reset every time an interruptFace() call is made. Usage:
# set of face poses
s = StringVec()
s.append("au_6_right")
# set of face values
v = FloatVec()
v.append(1.0)
fd.setDefaultFacePose(s, v)
brad.updateFaceDefinition()

+++ core/smartbody/SmartBody/src/controllers/me_ct_motion_graph.cpp
@@ -36,7 +36,7 @@ bool MeCtMotionGraph::controller_evaluate( double t, MeFrameData& frame )
 		outMotionBuffer.copyFrameBuffer(motionState->getCurMotionFrameBuffer());
 	}	
 	
-	updateDt(t);
+	updateDt((float) t);
 	if (dt > 0.05f)
 		dt = 0.05f; // cap the update rate
